+++ constants/constants.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int main()
 {
-	const int minutes_per_hour = 60;
+	const unsigned short minutes_per_hour = 60;
 	
 
 	cout << minutes_per_hour << endl;
