+++ homework_3.1/homework_3.1.cpp
@@ -4,14 +4,14 @@ using namespace std;
 
 int main()
 {
-	int age;
+	unsigned short age;
 
 
 	cout << "Please enter your age: ";
 	cin >> age;
 	cout << "\n";
 
-	int your_age_after_5_years = age + 5;
+	unsigned short your_age_after_5_years = age + 5;
 
 	cout << "After 5 years you will be " << your_age_after_5_years << " years old.\n";
